+++ KickIt/KickIt/Global/Helpers/DateExtensions.swift
@@ -42,3 +42,26 @@ func dateTimeToString(date3: Date) -> String {
     
     return datetimeToString
 }
+
+/// 경기 이벤트 발생 시간에 따라 실제 시간 계산하는 함수
+func setEventTime(plusMinute: Int) -> String {
+    var dateComponents = DateComponents()
+    dateComponents.year = 2024
+    dateComponents.month = 5
+    dateComponents.day = 18
+    dateComponents.hour = 16
+    dateComponents.minute = 00
+    dateComponents.second = 0
+    
+    // 기본 날짜 생성
+    var realTime = Calendar.current.date(from: dateComponents) ?? Date()
+    
+    // plusMinute를 realTime에 더함
+    if let updatedTime = Calendar.current.date(byAdding: .minute, value: plusMinute, to: realTime) {
+        realTime = updatedTime
+    }
+    
+    let realTimeString = dateTimeToString(date3: realTime)
+    
+    return realTimeString
+}
